Chat security

+++ backend/app/src/chat/chat.controller.ts
@@ -39,11 +39,15 @@ export class ChatController {
 		@Request() request: RequestWithUser,
 		@Body() body: { hidden: boolean, chatroom_id: number }
 	): void {
+		if (request.user.owner.indexOf(id) === -1)
+			throw new HttpException("User is not the chatroom owner", HttpStatus.FORBIDDEN);
 		this.chatService.setHidden(request.user.id, body.chatroom_id, body.hidden);
 	}
 
 	@Delete(':id')
 	delete(@Request() request: RequestWithUser, @Param('id') id: number): void {
+		if (request.user.owner.indexOf(id) === -1)
+			throw new HttpException("User is not the chatroom owner", HttpStatus.FORBIDDEN);
 		this.chatService.delete(request.user.id, id);
 	}
 */
@@ -61,6 +65,8 @@ export class ChatController {
 		@Request() request: RequestWithUser,
 		@Body() body: { admin: boolean, id: number, user_id: number }
 	): void {
+		if (request.user.owner.indexOf(id) === -1)
+			throw new HttpException("User is not the chatroom owner", HttpStatus.FORBIDDEN);
 		this.chatService.setAdmin(request.user.id, body.admin, body.id, body.user_id); // check if request.user.id is chatroom owner_uid
 	}
 
@@ -78,11 +84,15 @@ export class ChatController {
 		@Request() request: RequestWithUser,
 		@Body() body: { type: string, chatroom_id: number, target_id: number}
 	): void {
+		if (request.user.admin.indexOf(chatroom_id) === -1)
+			throw new HttpException("User is not the chatroom owner", HttpStatus.FORBIDDEN);
 		this.chatService.setPunishment(request.user.id, body.chatroom_id, body.target_id, body.type); // Check if request.user.id is admin
 	}
 
 	@Delete('kick/:id/:user_id')
 	kick(@Request() request: RequestWithUser, @Param('id') id: number, @Param('user_id') user_id: number): void {
+		if (request.user.admin.indexOf(chatroom_id) === -1)
+			throw new HttpException("User is not the chatroom owner", HttpStatus.FORBIDDEN);
 		this.chatService.kick(request.user.id, id, user_id); // Check if request.user.id is admin
 	}
 */
